requirements.txt
beautifulsoup4==4.12.3
loguru==0.7.2
lxml==5.2.2
openai==1.40.3
pandas==2.2.2
Pillow==10.4.0
pydantic_settings==2.4.0
Requests==2.32.3
xhs==0.2.13
